requirements.txt
asgiref==3.8.1
Django==4.2.10
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-paginator==0.2a3
pillow==10.3.0
sqlparse==0.5.0
